Add github workflow to build and push docker image
This commit introduces a simple github workflow. The workflow will
build healthy-api and healthy-ui images and push them to the docker
registry und hechi/healthy-[api/ui]. The workflow is only executed
on the master branch and will tag the image as *dev*. To indicate
that the image is the development image.