Fix loading environmen variables in react
Currently it is not possible to load the enviroment
variables after the build process. The only way is
to trick the nginx by generating an env-config.js
via env.sh which is loading the environment variables
into the file above. This file is later provided by
the nginx and loaded in the react application.

The solution is based on: https://www.freecodecamp.org/news/how-to-implement-runtime-environment-variables-with-create-react-app-docker-and-nginx-7f9d42a91d70/